python-app/requirements.txt
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp
